DI implementation completed, now also the interface of the messenger is passed as a parameter to the viewmodel and to the contract implementation. As a result now the application manager is a lot simpler (and there is no need to be a singleton because no classes shold call it).
Changed parameters in ThemManager: considering that the application is always the same, it's directly resolved inside the library method instead of being passed as a parameter.

NB: configuration's controller is broken, i mostly but not completely switched to an int base system (where the int is the id of the controller) instead of string base system (where the string was the name assigned to the controller). To test this i need to modify DS4Tester.exe and DS4Service.exe and will do it as a next and last change to the tester.