Overloaded ControllerContract: considering that it's used for different kind of messages, now it need only the parameter used for the occasion.
Add check of service during client startup: if the service is not running it fire a message and close itself.

Edit TranslationLibrary: if a string passed didn't have any translation, it return the string instead of a strange format string.

Add translation of EventMessage from the service: it's still in english on the service but when the client read it will try to translate before adding to the list of messages.